adding reference to player argument
