+++ ValheimPlus/GameClasses/Attack.cs
@@ -1,0 +1,62 @@
+﻿using HarmonyLib;
+using ValheimPlus.Configurations;
+
+namespace ValheimPlus
+{
+    /// <summary>
+    /// Alters stamina of weapons
+    /// </summary>
+    [HarmonyPatch(typeof(Attack), "GetStaminaUsage")]
+    public static class Attack_GetStaminaUsage_Patch
+    {
+        private static void Postfix(ref Attack __instance, ref float __result)
+        {
+            if (Configuration.Current.StaminaUsage.IsEnabled)
+            {
+                if (__instance.m_character.IsPlayer())
+                {
+                    ItemDrop.ItemData item = __instance.m_character.GetCurrentWeapon();
+                    Skills.SkillType skillType;
+                    if (item == null)
+                    {
+                        skillType = Skills.SkillType.Unarmed;
+                    }
+                    else
+                    {
+                        skillType = item.m_shared.m_skillType;
+                    }
+
+                    switch (skillType)
+                    {
+                        case Skills.SkillType.Swords:
+                            __result = Helper.applyModifierValue(__result, Configuration.Current.StaminaUsage.swords);
+                            break;
+                        case Skills.SkillType.Knives:
+                            __result = Helper.applyModifierValue(__result, Configuration.Current.StaminaUsage.knives);
+                            break;
+                        case Skills.SkillType.Clubs:
+                            __result = Helper.applyModifierValue(__result, Configuration.Current.StaminaUsage.clubs);
+                            break;
+                        case Skills.SkillType.Polearms:
+                            __result = Helper.applyModifierValue(__result, Configuration.Current.StaminaUsage.polearms);
+                            break;
+                        case Skills.SkillType.Spears:
+                            __result = Helper.applyModifierValue(__result, Configuration.Current.StaminaUsage.spears);
+                            break;
+                        case Skills.SkillType.Axes:
+                            __result = Helper.applyModifierValue(__result, Configuration.Current.StaminaUsage.axes);
+                            break;
+                        case Skills.SkillType.Unarmed:
+                            __result = Helper.applyModifierValue(__result, Configuration.Current.StaminaUsage.unarmed);
+                            break;
+                        case Skills.SkillType.Pickaxes:
+                            __result = Helper.applyModifierValue(__result, Configuration.Current.StaminaUsage.pickaxes);
+                            break;
+                        default:
+                            break;
+                    }
+                }
+            }
+        }
+    }
+}
